CKI-824 exclude k8s-api-server from istio
capi-kpack-watcher starts a watch for kpack "builds" on the api-server.
We suspect that istio/envoy terminates the connection after some
timeout-value and capi-kpack-watcher doesn't correctly reconnect.

As a workaround we have added an istio-annotation as an overlay to the
capi-kpack-watcher deployment to exclude "100.64.0.1/32" from istio
routing, which is the IP of the kubernetes-api-server service on
gardener with default CIDRs.

Co-authored-by: Philipp Stehle <philipp.stehle@sap.com>